refactor(spa-editor): share the data-type source-policy query
The Connections routing rows need the profile's raw DataTypeSourcePolicy
rows; `useSourcePolicies` already read them but immediately narrowed to the
types with 2+ enabled import sources. Extracting the live query gives both
callers one definition of what "no row for this type" means instead of two
copies of the same Dexie read drifting apart. No behaviour change.
